Begin importing Sequencescape consumers
This is pretty much a copy-paste of the Sequencescape code
followed by some re-organization to get specs passing without
ActiveRecord.

Eventually we are going to want to include Active Record for testing
but I really don't want to introduce it as a runtime dependency,
to need to be cautious about loading it.