Improve warren error recovery
`ActiveRecord::Base.connected?` merely check that a connection has been
checked out for the thread, not that its actually active. Unlike
`ActiveRecord::Base.connection.active?`. However there are additional
complications:

`ActiveRecord::Base.connection` can throw an exception if we don't yet
have a connection. So we need to rescue and return true. Unfortunately
the class of the exception depends on the database adaptor.

with_connection also throws an exception if it can't connect. However
our rescue was inside this block, so was missing it. However, if we
move it outside the block then there is no guarantee that we'll test
the same connection we tried to use.

So we handle checkout of the connection manually. We rescue all
exceptions, as we can be pretty sure the failure isn't a property
of the message, so re-queuing is appropriate.